backend/config/requirements.txt
# Web框架
fastapi>=0.118.0  # Updated for Python 3.13 compatibility
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库
sqlalchemy==2.0.23
psycopg[binary]>=3.1.0
alembic>=1.13.0

# 数据处理
pandas>=2.2.0  # Supports Python 3.13
pydantic>=2.12.0  # Updated for Python 3.13 compatibility

# PDF处理
pdfplumber==0.10.3
PyPDF2==3.0.1
PyMuPDF==1.23.8

# Office文档处理
python-docx==1.1.0
python-pptx>=0.6.21
openpyxl>=3.0.0  # Required for pandas.read_excel() to read .xlsx files

# 图像处理和OCR
Pillow>=10.0.0  # Updated for Python 3.13 compatibility
numpy>=2.0.0,<3.0.0  # numpy 2.x for Python 3.13 support
easyocr>=1.7.2  # Updated version with better compatibility
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python>=4.8.0  # Supports numpy 2.x and Python 3.13

# NLP处理
transformers>=4.56.0  # Updated for Python 3.13 compatibility
torch>=2.0.0,<2.5.0  # Supports numpy 2.x and Python 3.13

# LLM API
openai==1.65.5

# HTTP客户端
requests==2.32.4
httpx==0.28.1

# Google Cloud Storage (for RAG file persistence on Cloud Run)
google-cloud-storage>=2.0.0

# 认证和安全
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.2.0
python-jose[cryptography]>=3.3.0

# 时区处理
pytz==2024.1

# 其他工具
python-multipart==0.0.20
chardet==5.2.0
python-dotenv==1.0.0
psutil>=5.9.0  # Process and system utilities
opencc-python-reimplemented>=0.1.7  # Traditional/Simplified Chinese conversion

debugpy>=1.8.0

# RAG Evaluation (RAGAS framework)
ragas>=0.2.0
datasets>=2.14.0
langchain-openai>=0.1.0

